More tests for saturating arithmetic -- exhaustive sampling of [U]Int8 arguments.
